Properly handle NoFatChain secondary-option on file-entries
Small files typically have this set in order to indicate that the FAT
can not be used, which is also non-sensical with small files (since
they'd never require a second cluster chain in the first place, and,
therefore, no FAT lookups).

*However*, the specification doesn't rule-out the possibility of using
adjacent cluster-chains for larger files when it can be accomodated. We
tried with a larger filesystem and still could not find any instances of
this. That said, the specification also seems to imply that the clusters
with directory entries *might* alays be adjacent. This remains to be
tested.

As far as how we know when to stop reading directory entries, we rely on
the end-of-directory entry-type (0x0) to know when to stop. In other
words, multiple types of information allowing for multiple ways to make
decisions, and some fuzziness due to lack of definition for a couple of
rules/specs.

- Cluster enumeration and directory enumeration now returns the list of
  visited clusters and sectors (for verbosity and/or debugging).